Cleanup the structure of the cli package.
Move all flags into cli/flags
Move usage help into cli/usage.go
